--- src/_workspace/SettlementSummary/index.ts
@@ -1,5 +1,0 @@
-export { SettlementSummary } from './SettlementSummary';
-export type {
-  SettlementSummaryProps,
-  SettlementSummaryMember,
-} from './SettlementSummary';
