+++ src/pages/confirmStep/ui/SettlementSummary/index.ts
@@ -1,0 +1,2 @@
+export { SettlementSummary } from './SettlementSummary';
+export type { SettlementSummaryProps } from './SettlementSummary';
